:recycle: 컴포넌트명 가독성있게 수정 & type.ts로 수정
