+++ libs/shared/bg-gray-modal/type-bg-gray-modal.ts
@@ -1,0 +1,6 @@
+interface UiBgGrayModalProps {
+  closeButtonSize?: number
+  children: React.ReactNode
+}
+
+export type { UiBgGrayModalProps }
