+++ libs/shared/bg-gray-modal/ui/ui-bg-gray-modal/ui-bg-gray-modal.module.scss
