--- libs/shared/common-modal/data-access/data-access-modal-content-mock.module.scss
@@ -1,34 +1,0 @@
-@use '@/styles/utils';
-
-.title {
-  margin-bottom: 32px;
-  font-size: 28px;
-}
-
-.subTitle {
-  margin-bottom: 8px;
-  font-size: 16px;
-}
-
-.inputWrapper {
-  margin-bottom: 24px;
-  padding: 16px 20px;
-  background-color: utils.$white;
-}
-
-.input {
-  width: 100%;
-  font-size: 16px;
-}
-
-.input::placeholder {
-  color: utils.$gray-brt2;
-}
-
-.register {
-  display: flex;
-  margin: 0 auto;
-  padding: 14px 126.5px;
-  background-color: #ea3c12;
-  color: utils.$white;
-}
