--- libs/shared/common-modal/type-common-modal.ts
@@ -1,6 +1,0 @@
-interface UiCommonModalProps {
-  closeButtonSize?: number
-  children: React.ReactNode
-}
-
-export type { UiCommonModalProps }
