--- libs/shared/common-toast/type-common-toast.d.ts
@@ -1,11 +1,0 @@
-type AnimationStatus = 'fadeIn' | 'fadeOut'
-
-interface UiCommonToastProps {
-  status: AnimationStatus
-  children: string
-}
-
-interface CommonToastProps {
-  onShow: (isShow: boolean) => void
-  children: string
-}
