--- libs/shared/common-toast/ui/ui-common-toast/ui-common-toast.tsx
@@ -1,12 +1,0 @@
-import classnames from 'classnames/bind'
-
-import styles from './ui-common-toast.module.scss'
-
-const cx = classnames.bind(styles)
-
-export default function UiCommonToast({
-  status,
-  children,
-}: UiCommonToastProps) {
-  return <div className={cx('toastContent', status)}>{children}</div>
-}
