+++ libs/shared/detail-filter/data-access/data-access-location.ts
