+++ libs/shared/detail-filter/ui/ui-detail-filter/ui-detail-filter.module.scss
