+++ libs/shared/detail-filter/ui/ui-divider/ui-divider.tsx
