+++ libs/shared/detail-filter/ui/ui-location-container/ui-location-container.module.scss
