+++ libs/shared/detail-filter/ui/ui-selected-chip/ui-selected-chip.tsx
