+++ libs/shared/dialog/ui/ui-action-dialog/ui-action-dialog.module.scss
@@ -1,0 +1,28 @@
+@use '@/styles/utils';
+
+.dialogContainer {
+  @include utils.flexbox(column, initial, center);
+
+  position: fixed;
+  top: 50%;
+  left: 50%;
+  width: 298px;
+  padding: 24px;
+  transform: translate(-50%, -50%);
+  border-radius: 12px;
+  background-color: utils.$white;
+}
+
+.checkImage {
+  margin-bottom: 16px;
+}
+
+.description {
+  margin-bottom: 32px;
+  @include utils.font-style('body1');
+}
+
+.buttonContainer {
+  display: flex;
+  gap: 8px;
+}
