--- libs/shared/modal/ui/ui-action-modal/ui-action-modal.module.scss
@@ -1,28 +1,0 @@
-@use '@/styles/utils';
-
-.modalContainer {
-  @include utils.flexbox(column, initial, center);
-
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  width: 298px;
-  padding: 24px;
-  transform: translate(-50%, -50%);
-  border-radius: 12px;
-  background-color: utils.$white;
-}
-
-.checkImage {
-  margin-bottom: 16px;
-}
-
-.description {
-  margin-bottom: 32px;
-  @include utils.font-style('body1');
-}
-
-.buttonContainer {
-  display: flex;
-  gap: 8px;
-}
