--- libs/shared/modal/ui/ui-confirm-modal/ui-confirm-modal.module.scss
@@ -1,24 +1,0 @@
-@use '@/styles/utils';
-
-.modalContainer {
-  @include utils.flexbox(column, initial, center);
-
-  position: fixed;
-  top: 50%;
-  left: 50%;
-  width: 298px;
-  padding: 24px;
-  transform: translate(-50%, -50%);
-  border-radius: 12px;
-  background-color: utils.$white;
-}
-
-.exclamationImage {
-  margin-bottom: 16px;
-}
-
-.description {
-  @include utils.font-style('body1');
-
-  margin-bottom: 32px;
-}
