+++ libs/shared/notification-modal/ui/ui-notification-modal/ui-notification-modal.tsx
@@ -2,12 +2,12 @@ import classNames from 'classnames/bind'
 
 import Image from 'next/image'
 
-import UiCommonLayout from '@/libs/shared/common-layout/ui/ui-common-layout/ui-common-layout'
 import { UiNotificationModalProps } from '@/libs/shared/notification-modal/type-notification-modal'
 import {
   UiAcceptedNotificationItem,
   UiRejectedNotificationItem,
 } from '@/libs/shared/notification-modal/ui/ui-notification-item/ui-notification-item'
+import UiTitleContentLayout from '@/libs/shared/title-content-layout/ui/ui-title-content-layout/ui-title-content-layout'
 
 import styles from './ui-notification-modal.module.scss'
 
@@ -30,7 +30,7 @@ export default function UiNotificationModal({
       >
         <Image src="/images/close.svg" alt="닫기" width={24} height={24} />
       </button>
-      <UiCommonLayout
+      <UiTitleContentLayout
         title={`알림 ${itemList.length}개`}
         titleSize={20}
         gap={16}
@@ -54,7 +54,7 @@ export default function UiNotificationModal({
             ),
           )}
         </div>
-      </UiCommonLayout>
+      </UiTitleContentLayout>
     </div>
   )
 }
