+++ libs/shared/title-content-layout/type-title-content-layout.ts
@@ -1,0 +1,9 @@
+interface UiTitleContentLayoutProps {
+  title: string
+  titleAlign?: 'start' | 'center'
+  titleSize?: number
+  gap?: number
+  children: React.ReactNode
+}
+
+export type { UiTitleContentLayoutProps }
