+++ libs/shared/toast/ui/ui-toast/ui-toast.module.scss
